commit: Restructure jest folder without changing functionality
