+++ framework/test/jest/config/jest.config.base.js
@@ -2,7 +2,7 @@ module.exports = {
 	verbose: true,
 	collectCoverage: true,
 	coverageReporters: ['lcov', 'cobertura'],
-	rootDir: '../../../../../',
+	rootDir: '../../../../',
 	setupFilesAfterEnv: ['<rootDir>/framework/test/jest/config/setup.js'],
 	/**
 		coverageThreshold: {
