+++ framework/test/jest/functional/jest.config.js
@@ -1,0 +1,9 @@
+const base = require('../config/jest.config.base');
+
+module.exports = {
+	...base,
+	testMatch: [
+		'<rootDir>/framework/test/jest/functional/specs/**/*.(spec|test).js',
+	],
+	coverageDirectory: '.coverage/functional',
+};
