+++ framework/test/jest/unit/specs/controller/channels/child_process_channel.spec.js
