+++ framework/test/jest/unit/specs/jest.config/jest.config.unit.spec.js
@@ -1,4 +1,4 @@
-const config = require('../../../config/unit/jest.config');
+const config = require('../../jest.config');
 
 describe('config/unit/jest.config.js', () => {
 	it('unit test config must match to the snapshot.', () => {
